Split out Firebase code into a signin component.
